requirements.txt
altgraph>=0.17.3
auto-py-to-exe>=2.34.0
Babel>=2.12.1
bottle>=0.12.25
bottle-websocket>=0.2.9
certifi>=2022.12.7
cffi>=1.15.1
charset-normalizer>=3.1.0
colorama>=0.4.6
cryptography>=39.0.2
Deprecated>=1.2.13
docx2pdf>=0.1.8
docxcompose>=1.4.0
docxtpl>=0.16.6
Eel>=0.16.0
future>=0.18.3
gcloud>=0.18.3
gevent>=22.10.2
gevent-websocket>=0.10.1
googleapis-common-protos>=1.58.0
greenlet>=2.0.2
httplib2>=0.21.0
idna>=3.4
Jinja2>=3.1.2
jwcrypto>=1.4.2
lxml>=4.9.2
MarkupSafe>=2.1.2
numpy>=1.24.2
oauth2client>=4.1.3
opencv-python>=4.7.0.72
pefile>=2023.2.7
Pillow>=9.4.0
protobuf>=4.22.1
pyasn1>=0.4.8
pyasn1-modules>=0.2.8
pycparser>=2.21
pycryptodome>=3.17
pyinstaller>=5.10.1
pyinstaller-hooks-contrib>=2023.2
pyparsing>=3.0.9
PyPDF2>=3.0.1
pypng>=0.20220715.0
Pyrebase4>=4.6.0
PySimpleGUI>=4.60.4
python-docx>=0.8.11
python-jwt>=4.0.0
pywin32>=306
pywin32-ctypes>=0.2.0
qrcode>=7.4.2
requests>=2.28.2
requests-toolbelt>=0.10.1
rsa>=4.9
schedule>=1.1.0
six>=1.16.0
tqdm>=4.65.0
typing_extensions>=4.5.0
urllib3>=1.26.14
whichcraft>=0.6.1
wrapt>=1.15.0
zope.event>=4.6
zope.interface>=6.0
